projects/toua-road/verification/phase3-agent/react_agent/requirements.txt
azure-functions
azure-search-documents>=11.6.0
azure-storage-blob>=12.19.0
azure-core
openai>=1.0.0
